Restructure E2E tests to share some common code
